feat: refactor CSFloat client integration and improve LLM service structure
