+++ backend/api/item_names.py
@@ -1,7 +1,6 @@
 from fastapi import APIRouter, HTTPException, Query
 from pydantic import BaseModel
-
-from ..services.csfloat_client import fetch_csfloat_item_names
+from services.csfloat.client import CSFloatClient
 
 
 class ItemNamesResponse(BaseModel):
@@ -9,12 +8,13 @@ class ItemNamesResponse(BaseModel):
 
 
 router = APIRouter()
+csfloat_client = CSFloatClient()
 
 
 @router.get("/item-names", response_model=ItemNamesResponse)
 def get_item_names(limit: int = Query(50, ge=1, le=500)):
     try:
-        names = fetch_csfloat_item_names(limit=limit)
+        names = csfloat_client.fetch_item_names(limit=limit)
         return {"names": names}
     except Exception as e:
         raise HTTPException(status_code=503, detail=f"Item names service unavailable: {str(e)}")
