+++ backend/api/llm_models.py
@@ -2,8 +2,7 @@ from typing import List
 
 from fastapi import APIRouter
 from pydantic import BaseModel
-
-from ..services.llm_registry import LLMModelInfo, list_models
+from services.llm.registry import LLMModelInfo, list_models
 
 router = APIRouter()
