+++ backend/services/llm/client.py
@@ -1,0 +1,6 @@
+from .service import ask_about_listings
+
+
+class LLMClient:
+    def ask_about_listings(self, *args, **kwargs):
+        return ask_about_listings(*args, **kwargs)
