--- backend/services/llm_client.py
@@ -1,3 +1,0 @@
-from .llm.service import ask_about_listings
-
-__all__ = ["ask_about_listings"]
