--- backend/services/llm_registry.py
@@ -1,3 +1,0 @@
-from .llm.registry import LLMModelInfo, list_models
-
-__all__ = ["LLMModelInfo", "list_models"]
